+++ src/IntlTimezone.ts
@@ -1,3 +1,4 @@
+import { OffsetTimezone } from './OffsetTimezone'
 import { Timezone } from './Timezone'
 import { tzUtc } from './UTCTimezone'
 import { MILLISECONDS_IN_MINUTE } from './constants'
@@ -47,115 +48,30 @@ function diffMinutes(d1: DateParts, d2: DateParts): number {
   return 60 * (24 * day + hour) + min
 }
 
-function getTimezoneOffset(date: Date, timeZone: string) {
-  const formatOptions: Intl.DateTimeFormatOptions = {
-    ...utcDateTimeFormatOptions,
-    timeZone,
-  }
+export class IntlTimezone extends OffsetTimezone {
+  private formatDateTz: Intl.DateTimeFormat
 
-  var formatDateTz = new Intl.DateTimeFormat(localeUS, formatOptions)
-
-  return -diffMinutes(
-    parseDate(formatDateUTC.format(date)),
-    parseDate(formatDateTz.format(date))
-  )
-}
-
-export class IntlTimezone extends Timezone {
   /**
-   * Construct a custom timezone.
+   * Construct a timezone which uses the browser Intl.DateTimeFormat
+   * to discover timezone offsets.
    *
-   * @param name The timezone name.
+   * @param timeZone The timezone name.
    * @param deltas The timezone offsets
    */
   constructor(name: string) {
     super(name)
-  }
 
-  /** @ignore */
-  #fromLocal(localDate: Date): Date {
-    const utcDate = new Date(
-      localDate.getTime() - this.offset(localDate) * MILLISECONDS_IN_MINUTE
-    )
-    return utcDate
-  }
-
-  /** @ignore */
-  #toLocal(utcDate: Date): Date {
-    const localDate = new Date(
-      utcDate.getTime() + this.offset(utcDate) * MILLISECONDS_IN_MINUTE
-    )
-    return localDate
-  }
-
-  makeDate(
-    year: number,
-    monthIndex: number,
-    day?: number,
-    hours?: number,
-    minutes?: number,
-    seconds?: number,
-    milliseconds?: number
-  ): Date {
-    const date = tzUtc.makeDate(
-      year,
-      monthIndex,
-      day,
-      hours,
-      minutes,
-      seconds,
-      milliseconds
-    )
-    return this.#fromLocal(date)
-  }
-
-  dateParts(date: Date): DatePartResponse {
-    const local = this.#toLocal(date)
-    return tzUtc.dateParts(local)
+    this.formatDateTz = new Intl.DateTimeFormat(localeUS, {
+      ...utcDateTimeFormatOptions,
+      timeZone: name,
+    })
   }
 
   offset(date: Date): number {
-    return getTimezoneOffset(date, this.name)
-  }
-
-  year(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.year(local)
-  }
-
-  monthIndex(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.monthIndex(local)
-  }
-
-  weekday(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.weekday(local)
-  }
-
-  day(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.day(local)
-  }
-
-  hours(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.hours(local)
-  }
-
-  minutes(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.minutes(local)
-  }
-
-  seconds(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.seconds(local)
-  }
-
-  milliseconds(date: Date): number {
-    const local = this.#toLocal(date)
-    return tzUtc.milliseconds(local)
+    return -diffMinutes(
+      parseDate(formatDateUTC.format(date)),
+      parseDate(this.formatDateTz.format(date))
+    )
   }
 
   isDaylightSavings(date: Date): boolean {
